chore: (#521) eventService 삭제
