--- dms-core/src/main/kotlin/team/aliens/dms/common/service/event/EventService.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.common.service.event
-
-import team.aliens.dms.domain.notification.model.Notification
-
-interface EventService {
-    fun publishNotification(deviceToken: String, notification: Notification)
-    fun publishNotificationToAll(notification: Notification)
-}
