+++ dms-core/src/main/kotlin/team/aliens/dms/domain/notification/service/NotificationService.kt
@@ -18,4 +18,8 @@ interface NotificationService {
     fun subscribeTopic(userId: UUID, topic: Topic)
 
     fun updateSubscribes(userId: UUID, topicSubscribes: List<Pair<Topic, Boolean>>)
+
+    fun publishNotification(deviceToken: String, notification: Notification)
+
+    fun publishNotificationToAll(notification: Notification)
 }
