+++ dms-core/src/main/kotlin/team/aliens/dms/domain/notification/service/NotificationServiceImpl.kt
@@ -2,6 +2,7 @@ package team.aliens.dms.domain.notification.service
 
 import java.util.UUID
 import org.springframework.stereotype.Component
+import team.aliens.dms.common.spi.EventPort
 import team.aliens.dms.domain.notification.exception.DeviceTokenNotFoundException
 import team.aliens.dms.domain.notification.model.DeviceToken
 import team.aliens.dms.domain.notification.model.Notification
@@ -12,7 +13,8 @@ import team.aliens.dms.domain.notification.spi.NotificationPort
 @Component
 class NotificationServiceImpl(
     private val notificationPort: NotificationPort,
-    private val deviceTokenPort: DeviceTokenPort
+    private val deviceTokenPort: DeviceTokenPort,
+    private val eventPort: EventPort
 ) : NotificationService {
 
     override fun saveDeviceToken(deviceToken: DeviceToken) {
@@ -35,32 +37,32 @@ class NotificationServiceImpl(
     }
 
     override fun subscribeTopic(userId: UUID, topic: Topic) {
-        val notificationToken = this.getDeviceTokenByUserId(userId)
+        val deviceToken = this.getDeviceTokenByUserId(userId)
         notificationPort.subscribeTopic(
-            deviceToken = notificationToken.deviceToken,
+            deviceToken = deviceToken.deviceToken,
             topic = topic
         )
     }
 
     override fun unsubscribeTopic(userId: UUID, topic: Topic) {
-        val notificationToken = this.getDeviceTokenByUserId(userId)
+        val deviceToken = this.getDeviceTokenByUserId(userId)
         notificationPort.unsubscribeTopic(
-            deviceToken = notificationToken.deviceToken,
+            deviceToken = deviceToken.deviceToken,
             topic = topic
         )
     }
 
     override fun updateSubscribes(userId: UUID, topicSubscribes: List<Pair<Topic, Boolean>>) {
-        val notificationToken = this.getDeviceTokenByUserId(userId)
+        val deviceToken = this.getDeviceTokenByUserId(userId)
         topicSubscribes.forEach { (topic, isSubscribe) ->
             if (isSubscribe) {
                 notificationPort.subscribeTopic(
-                    deviceToken = notificationToken.deviceToken,
+                    deviceToken = deviceToken.deviceToken,
                     topic = topic
                 )
             } else {
                 notificationPort.unsubscribeTopic(
-                    deviceToken = notificationToken.deviceToken,
+                    deviceToken = deviceToken.deviceToken,
                     topic = topic
                 )
             }
@@ -69,4 +71,22 @@ class NotificationServiceImpl(
 
     private fun getDeviceTokenByUserId(userId: UUID) =
         deviceTokenPort.queryDeviceTokenById(userId) ?: throw DeviceTokenNotFoundException
+
+    override fun publishNotification(
+        deviceToken: String,
+        notification: Notification
+    ) {
+        eventPort.publishNotification(
+            deviceToken = deviceToken,
+            notification = notification
+        )
+    }
+
+    override fun publishNotificationToAll(
+        notification: Notification
+    ) {
+        eventPort.publishNotificationToAll(
+            notification = notification
+        )
+    }
 }
